Add reusable Google sign-in helper
